PROJECT CREATION WITH MAVEN
Two enumareations have been created, one for the agencies and one for
different image types.

The Agency Class is an abstract class that provides the getters for all
the required fields of each agency. Each agency will have its own class
to return the requirment or detail (such as FRP url).

The Agency Factory will be used to return a class of AgencyDetails to
eliminate the initialization of the objects inside the other classes.

ShuttestockDetails has been created which inherits from AgencyDetails
however not all details have been added.

EPS, JPEG and TIFF files are added for the moment, video will be added
later but the extensions should be small.

After ShutterstockDetails has been completed a test class will be
written for the ShutterstockDetails. Also a tester for the AgencyFactory