CREATED CLASS CONNECT TO DB + CHECK TABLE
Wrote a function to connect to SLQlite DB - tried to make it cross
platform. Testing needed on other OS.

Wrote function to connect to DB and check if table for credentials
exists. Now need to write function to create the table, perhaps this
will be moved to the DBFactory/Manager class