ADDED STATE VARIABLE TO KEEP THE STATE FOR OBSERVERS
Main addition is the state variable, which updates the observers
whenever the total percentage of update changes by 0.1%.

Notify method can have as parameter the agency issuing the change of
state, to only update single agency details on the gui. This would
make changes to gui less frequent (per agency).